* no need for Uint64 when SDL_GetTicks only returns 32bit unsigned
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@4588 37977431-3df6-0310-b722-df95706aa16b

+++ rts/Game/Game.h
@@ -61,14 +61,14 @@ public:
 	unsigned int thisFps;
 
 	time_t fpstimer, starttime;
-	Uint64 lastUpdate;
-	Uint64 lastMoveUpdate;
-	Uint64 lastModGameTimeMeasure;
+	unsigned lastUpdate;
+	unsigned lastMoveUpdate;
+	unsigned lastModGameTimeMeasure;
 
-	Uint64 lastUpdateRaw;
+	unsigned lastUpdateRaw;
 	float updateDeltaSeconds;
 
-	unsigned int lastframe;
+	unsigned lastframe;
 	float totalGameTime;			//time in seconds, stops at game end
 
 	std::string userInputPrefix;
@@ -139,7 +139,7 @@ protected:
 	float inputTextSizeY;
 	float lastCpuUsageTime;
 	
-	Uint32 lastFrameTime;
+	unsigned lastFrameTime;
 };
